refactor - remove datagen from public deployment
