+++ index.html
@@ -7,7 +7,7 @@
     <title>Vite App</title>
   </head>
   <body>
-    <div id="app" class="sm:w-screen m-auto"></div>
+    <div id="app" class="sm:w-screen"></div>
     <script type="module" src="/src/main.js"></script>
     <section
     class="bg-black flex min-h-fit justify-around items-center flex-wrap max-h-screen"
