+++ src/index.css
@@ -2,6 +2,8 @@
 @tailwind components;
 @tailwind utilities;
 
-#app{
-    width: fit-content;
+#app {
+  width: fit-content;
+  border-left: auto;
+  border-right: auto;
 }
